ge: graphics: rename Renderer Factory to Graphics Factory
