+++ src/genesis/graphics/CMakeLists.txt
@@ -17,9 +17,9 @@ list(APPEND GE_RENDERER_HEADERS
     ${GE_RENDERER_INCLUDE_DIR}/gpu_command_queue.h
     ${GE_RENDERER_INCLUDE_DIR}/graphics.h
     ${GE_RENDERER_INCLUDE_DIR}/graphics_context.h
+    ${GE_RENDERER_INCLUDE_DIR}/graphics_factory.h
     ${GE_RENDERER_INCLUDE_DIR}/index_buffer.h
     ${GE_RENDERER_INCLUDE_DIR}/render_command.h
-    ${GE_RENDERER_INCLUDE_DIR}/renderer_factory.h
     ${GE_RENDERER_INCLUDE_DIR}/shader.h
     ${GE_RENDERER_INCLUDE_DIR}/shader_input_layout.h
     ${GE_RENDERER_INCLUDE_DIR}/shader_precompiler.h
